+++ src/color/gradient.rs
@@ -1,3 +1,4 @@
+use crate::color::Color;
 use egui::color::{Color32, Rgba};
 use egui::lerp;
 
@@ -6,6 +7,10 @@ pub struct Gradient(pub Vec<Color32>);
 
 #[allow(dead_code)]
 impl Gradient {
+    pub fn from_colors(colors: impl IntoIterator<Item = Color>) -> Self {
+        Self(colors.into_iter().map(|c| c.color32()).collect())
+    }
+
     pub fn one_color(srgba: Color32) -> Self {
         Self(vec![srgba, srgba])
     }
@@ -16,22 +21,19 @@ impl Gradient {
             .map(|color| format!("{:02x}{:02x}{:02x}", color.r(), color.g(), color.b()))
     }
 
-    pub fn texture_gradient(left: Color32, right: Color32) -> Self {
-        Self(vec![left, right])
+    pub fn texture_gradient(left: Color, right: Color) -> Self {
+        Self(vec![left.color32(), right.color32()])
     }
-    pub fn ground_truth_linear_gradient(left: Color32, right: Color32) -> Self {
-        let left = Rgba::from(left);
-        let right = Rgba::from(right);
+    pub fn ground_truth_linear_gradient(left: Color, right: Color) -> Self {
+        let left = left.rgba();
+        let right = right.rgba();
 
-        let n = 255;
-        Self(
-            (0..=n)
-                .map(|i| {
-                    let t = i as f32 / n as f32;
-                    Color32::from(lerp(left..=right, t))
-                })
-                .collect(),
-        )
+        const N: f32 = 255.0;
+        let mut colors = vec![];
+        for i in 0..=N as i32 {
+            colors.push(Color32::from(lerp(left..=right, i as f32 / N)));
+        }
+        Self(colors)
     }
 
     /// Do premultiplied alpha-aware blending of the gradient on top of the fill color