+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add a checkbox to disable color labels in harmonies view
 - Fix double `#` when copying a list of hex colors
 - Add different harmony layouts like `square`, `line` or `stacked`
+- Add gradient view of harmony
 
 # 0.5.1
 - Fix switching between working spaces and illuminants